refactor: extract residual rendering into useResidualPlot hook
Move the Residual Plot's SVG rendering, split-layout sync autorun, and
selection-halo restyle out of scatter-plot.tsx into a useResidualPlot hook,
slimming the core component. Behavior-preserving; existing residual tests green.
